Add files via upload
## COVID-19 Global Dataset

### Description:
This dataset provides a comprehensive overview of the global impact of the COVID-19 pandemic. It includes data on confirmed cases, recoveries, deaths, and vaccination rates across different countries and regions. The data is regularly updated and is sourced from reliable public health organizations.

### Data Columns:
- Country/Region: The country or region where the data is recorded.
- Date: The date on which the data was recorded.
- Confirmed Cases: The cumulative number of confirmed COVID-19 cases.
- Deaths:The cumulative number of deaths due to COVID-19.
- Recovered: The cumulative number of recoveries from COVID-19.
- active Cases: The number of currently active cases.
- Vaccination Rate: The percentage of the population that has received at least one dose of the COVID-19 vaccine.

### Sources:
The dataset is aggregated from the following sources:
- World Health Organization (WHO)
- Johns Hopkins University Center for Systems Science and Engineering (JHU CSSE)
- Local health departments and government websites

### Usage:
This dataset is ideal for creating visualizations, conducting trend analysis, or studying the impact of COVID-19 on a global scale. Feel free to use this data in your projects, research, or for educational purposes.

### Contributing:
If you would like to contribute to this dataset, feel free to submit a pull request. Contributions such as data updates, corrections, and enhancements are welcome.